requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
emoji==0.5.3
ephem==3.7.6.0
future==0.17.1
pycparser==2.19
PySocks==1.7.0
python-telegram-bot[SOCKS]==11.1.0
six==1.12.0
clarifai==2.2.3
